added checkout page and moved post functionality here with total calculation
