+++ bike-rental/views/bikeDetails.ejs
@@ -12,7 +12,7 @@
         <a href="/user">User Profile</a>
       </nav>
     </header>
-    <form action="/bikeDetails" method="post">
+    <form action="/checkout" method="get">
       <main>
         <h1>Bike Details</h1>
   
@@ -34,17 +34,11 @@
       <input type="date" id="end-date" name="end-date"><br>
     </section>
 
-    <section class="checkout-section">
-      <h1>Checkout Details</h1>
-      <p>Total: </p>
-    </section>
-
       <input type="submit" value="Rent Bike" class="button">
       </form>
 
-      <a href="/bikeRental" class="back-button">Back</a>
+      <a href="/bikeRental" class="back-button">Back</a>   
   </body>
-  <p> hello </p>
 
 
 </html>
